android-14.0.0_r2 part 5

+++ system/chre/core/ble_request.cc
@@ -174,14 +174,14 @@ bool BleRequest::isEnabled() const {
 void BleRequest::logStateToBuffer(DebugDumpWrapper &debugDump,
                                   bool isPlatformRequest) const {
   if (!isPlatformRequest) {
-    debugDump.print("  instanceId=%" PRIu32 " status=%" PRIu8, mInstanceId,
+    debugDump.print("  instanceId=%" PRIu16 " status=%" PRIu8, mInstanceId,
                     static_cast<uint8_t>(mStatus));
   }
   debugDump.print(" %s", mEnabled ? " enable" : " disable\n");
   if (mEnabled) {
-    debugDump.print(" mode=%" PRIu8 " reportDelayMs=%" PRIu32
-                    " rssiThreshold=%" PRId8,
-                    mMode, mReportDelayMs, mRssiThreshold);
+    debugDump.print(
+        " mode=%" PRIu8 " reportDelayMs=%" PRIu32 " rssiThreshold=%" PRId8,
+        static_cast<uint8_t>(mMode), mReportDelayMs, mRssiThreshold);
     if (isPlatformRequest) {
       debugDump.print(" filters=[");
       for (const chreBleGenericFilter &filter : mFilters) {
